pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "long-context-bench"
version = "0.1.0"
description = "A benchmark for evaluating long-context code editing capabilities of CLI-based coding agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Augment Code", email = "aj@augmentcode.com"}
]
keywords = ["benchmark", "code-editing", "long-context", "agents", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "gitpython>=3.1.40",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
    "tenacity>=8.2.3",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
]

llm-judge = [
    "openai>=1.6.0",
    "anthropic>=0.8.0",
]

[project.scripts]
long-context-bench = "long_context_bench.cli:main"

[project.urls]
Homepage = "https://github.com/AugmentedAJ/Long-Context-Code-Bench"
Repository = "https://github.com/AugmentedAJ/Long-Context-Code-Bench"
Issues = "https://github.com/AugmentedAJ/Long-Context-Code-Bench/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=long_context_bench --cov-report=term-missing"

